Wrapping LHCb functions, moving wrapper definition
